+++ src/Finance.Core/Processes/ProcessHoursEnteritem.cs
@@ -31,7 +31,7 @@ namespace Finance.Processes
 		/// 
 		/// </summary>
 		[Column("process_hours_enter_id")]
-		public decimal? ProcessHoursEnterId { get; set; }
+		public long? ProcessHoursEnterId { get; set; }
 
 		/// <summary>
 		/// 年
